refactor: cleanup tests and move getExportSubpathConfigs to utils
