+++ tools/workspace-plugin/src/executors/generate-api/lib/shared.ts
@@ -1,7 +1,15 @@
+import { logger } from '@nx/devkit';
+
 export function isCI() {
   return (
     (process.env.CI && process.env.CI !== 'false') ||
     (process.env.TF_BUILD && process.env.TF_BUILD.toLowerCase() === 'true') ||
     process.env.GITHUB_ACTIONS === 'true'
   );
+}
+
+export function verboseLog(message: string, kind: keyof typeof logger = 'info') {
+  if (process.env.NX_VERBOSE_LOGGING === 'true') {
+    logger[kind](message);
+  }
 }
